Change test for dealing with long long int.

+++ tst/dtkCore/dtkHashTest.cpp
@@ -44,16 +44,16 @@ void dtkHashTestCase::init(void)
 
 void dtkHashTestCase::testPair(void)
 {
-    std::unordered_map<std::pair<int, int>, int, dtkHash> map_of_pair;
+    std::unordered_map<std::pair<qlonglong, qlonglong>, qlonglong, dtkHash> map_of_pair;
 
-    std::pair<int, int> p0 = std::make_pair(12, 19);
-    std::pair<int, int> p1 = std::make_pair(19, 12);
+    std::pair<qlonglong, qlonglong> p0 = std::make_pair(12, 19);
+    std::pair<qlonglong, qlonglong> p1 = std::make_pair(19, 12);
 
     map_of_pair[p0] =  27;
     map_of_pair[p1] = -27;
 
-    QCOMPARE(map_of_pair[p0],  27);
-    QCOMPARE(map_of_pair[p1], -27);
+    QCOMPARE(map_of_pair[p0],  27LL);
+    QCOMPARE(map_of_pair[p1], -27LL);
 }
 
 void dtkHashTestCase::cleanup(void)
